+++ src/memory.cpp
@@ -81,7 +81,7 @@ address_class_t::get_info (amd_dbgapi_address_class_info_t query,
 
 bool
 address_space_t::address_is_in_address_class (
-  const wave_t & /* wave */, amd_dbgapi_lane_id_t /* lane_id  */,
+  const wave_t &wave, amd_dbgapi_lane_id_t /* lane_id */,
   amd_dbgapi_segment_address_t segment_address,
   const address_class_t &address_class) const
 {
@@ -106,7 +106,7 @@ address_space_t::address_is_in_address_class (
 
   /* The private_swizzled address space is in private address class, the local
      address space is in the local address class, etc...  */
-  return lower (segment_address).first.kind ()
+  return lower (wave.agent (), segment_address).first.kind ()
          == address_class.address_space ().kind ();
 }
 
@@ -169,6 +169,7 @@ address_space_t::host ()
 
 std::pair<const address_space_t &, amd_dbgapi_segment_address_t>
 global_address_space_t::lower (
+  const agent_t & /* agent */,
   amd_dbgapi_segment_address_t global_address) const
 {
   return { *this, utils::zero_extend (global_address, address_size ()) };
@@ -181,7 +182,7 @@ global_address_space_t::convert (
   amd_dbgapi_segment_address_t from_address) const
 {
   auto [lowered_address_space, lowered_address]
-    = from_address_space.lower (from_address);
+    = from_address_space.lower (wave.agent (), from_address);
 
   if (lowered_address == lowered_address_space.null_address ()
       && (from_address_space.kind () == kind_t::generic
@@ -235,19 +236,20 @@ global_address_space_t::convert (
 }
 
 std::pair<const address_space_t &, amd_dbgapi_segment_address_t>
-local_address_space_t::lower (amd_dbgapi_segment_address_t local_address) const
+local_address_space_t::lower (const agent_t & /* agent */,
+                              amd_dbgapi_segment_address_t local_address) const
 {
   return { *this, utils::zero_extend (local_address, address_size ()) };
 }
 
 std::pair<amd_dbgapi_segment_address_t, amd_dbgapi_size_t>
 local_address_space_t::convert (
-  const wave_t & /* wave  */, amd_dbgapi_lane_id_t /* lane_id  */,
+  const wave_t &wave, amd_dbgapi_lane_id_t /* lane_id  */,
   const address_space_t &from_address_space,
   amd_dbgapi_segment_address_t from_address) const
 {
   auto [lowered_address_space, lowered_address]
-    = from_address_space.lower (from_address);
+    = from_address_space.lower (wave.agent (), from_address);
 
   if (lowered_address == lowered_address_space.null_address ()
       && (from_address_space.kind () == kind_t::generic
@@ -262,6 +264,7 @@ local_address_space_t::convert (
 
 std::pair<const address_space_t &, amd_dbgapi_segment_address_t>
 private_swizzled_address_space_t::lower (
+  const agent_t & /* agent */,
   amd_dbgapi_segment_address_t private_address) const
 {
   return { *this, utils::zero_extend (private_address, address_size ()) };
@@ -274,7 +277,7 @@ private_swizzled_address_space_t::convert (
   amd_dbgapi_segment_address_t from_address) const
 {
   auto [lowered_address_space, lowered_address]
-    = from_address_space.lower (from_address);
+    = from_address_space.lower (wave.agent (), from_address);
 
   if (lowered_address == lowered_address_space.null_address ()
       && (from_address_space.kind () == kind_t::generic
@@ -321,6 +324,7 @@ private_swizzled_address_space_t::convert (
 
 std::pair<const address_space_t &, amd_dbgapi_segment_address_t>
 private_unswizzled_address_space_t::lower (
+  const agent_t & /* agent */,
   amd_dbgapi_segment_address_t private_address) const
 {
   return { *this, utils::zero_extend (private_address, address_size ()) };
@@ -333,7 +337,7 @@ private_unswizzled_address_space_t::convert (
   amd_dbgapi_segment_address_t from_address) const
 {
   auto [lowered_address_space, lowered_address]
-    = from_address_space.lower (from_address);
+    = from_address_space.lower (wave.agent (), from_address);
 
   if (lowered_address == lowered_address_space.null_address ()
       && (from_address_space.kind () == kind_t::generic
@@ -407,7 +411,7 @@ generic_address_space_t::generic_address_space_t (
 
 std::optional<amd_dbgapi_segment_address_t>
 generic_address_space_t::generic_address_for_address_space (
-  const address_space_t &segment_address_space,
+  const agent_t & /* agent */, const address_space_t &segment_address_space,
   amd_dbgapi_segment_address_t segment_address) const
 {
   for (auto &&aperture : m_apertures)
@@ -417,7 +421,7 @@ generic_address_space_t::generic_address_for_address_space (
           return null_address ();
 
         return aperture.base
-               | utils::zero_extend (segment_address,
+               + utils::zero_extend (segment_address,
                                      segment_address_space.address_size ());
       }
 
@@ -425,16 +429,9 @@ generic_address_space_t::generic_address_for_address_space (
   return std::nullopt;
 }
 
-amd_dbgapi_segment_address_dependency_t
-generic_address_space_t::address_dependency (
-  amd_dbgapi_segment_address_t address) const
-{
-  auto [lowered_address_space, lowered_address] = lower (address);
-  return lowered_address_space.address_dependency (lowered_address);
-}
-
 std::pair<const address_space_t &, amd_dbgapi_segment_address_t>
 generic_address_space_t::lower (
+  const agent_t & /* agent */,
   amd_dbgapi_segment_address_t generic_address) const
 {
   for (auto &&aperture : m_apertures)
@@ -464,10 +461,10 @@ generic_address_space_t::convert (
   amd_dbgapi_segment_address_t from_address) const
 {
   auto [lowered_address_space, lowered_address]
-    = from_address_space.lower (from_address);
+    = from_address_space.lower (wave.agent (), from_address);
 
   auto generic_address = generic_address_for_address_space (
-    lowered_address_space, lowered_address);
+    wave.agent (), lowered_address_space, lowered_address);
 
   if (!generic_address)
     throw api_error_t (
@@ -498,21 +495,37 @@ generic_address_space_t::convert (
 }
 
 std::pair<const address_space_t &, amd_dbgapi_segment_address_t>
-agent_address_space_t::lower (amd_dbgapi_segment_address_t agent_address) const
+agent_address_space_t::lower (const agent_t & /* agent  */,
+                              amd_dbgapi_segment_address_t agent_address) const
 {
   return { *this, agent_address };
 }
 
 std::pair<amd_dbgapi_segment_address_t, amd_dbgapi_size_t>
 agent_address_space_t::convert (
-  const wave_t & /* wave  */, amd_dbgapi_lane_id_t /* lane_id  */,
+  const wave_t &wave, amd_dbgapi_lane_id_t /* lane_id  */,
   const address_space_t &from_address_space,
   amd_dbgapi_segment_address_t from_address) const
 {
   auto [lowered_address_space, lowered_address]
-    = from_address_space.lower (from_address);
+    = from_address_space.lower (wave.agent (), from_address);
 
   if (lowered_address_space.kind () == kind_t::agent)
+    return { lowered_address, last_address () - lowered_address + 1 };
+
+  throw api_error_t (AMD_DBGAPI_STATUS_ERROR_INVALID_ADDRESS_SPACE_CONVERSION);
+}
+
+std::pair<amd_dbgapi_segment_address_t, amd_dbgapi_size_t>
+host_address_space_t::convert (const wave_t &wave,
+                               amd_dbgapi_lane_id_t /* lane_id  */,
+                               const address_space_t &from_address_space,
+                               amd_dbgapi_segment_address_t from_address) const
+{
+  auto [lowered_address_space, lowered_address]
+    = from_address_space.lower (wave.agent (), from_address);
+
+  if (lowered_address_space.kind () == kind_t::host)
     return { lowered_address, last_address () - lowered_address + 1 };
 
   throw api_error_t (AMD_DBGAPI_STATUS_ERROR_INVALID_ADDRESS_SPACE_CONVERSION);
@@ -1086,18 +1099,15 @@ amd_dbgapi_convert_address_space (
     else if (lane_id != AMD_DBGAPI_LANE_NONE)
       THROW (AMD_DBGAPI_STATUS_ERROR_INVALID_LANE_ID);
 
-    auto [lowered_address_space, lowered_address]
-      = source_address_space->lower (source_segment_address);
-
     /* Handle global->global conversions early since it does not require to
        pass in a wave_id.  */
-    if (lowered_address_space.kind () == address_space_t::kind_t::global
+    if (source_address_space->kind () == address_space_t::kind_t::global
         && destination_address_space->kind ()
              == address_space_t::kind_t::global)
       {
-        *destination_segment_address = lowered_address;
+        *destination_segment_address = source_segment_address;
         *destination_contiguous_bytes
-          = lowered_address_space.last_address () - lowered_address + 1;
+          = source_address_space->last_address () - source_segment_address + 1;
       }
     else
       {
@@ -1108,7 +1118,7 @@ amd_dbgapi_convert_address_space (
            converion requires it (throws STATUS_ERROR_INVALID_LANE_ID).  */
         std::tie (*destination_segment_address, *destination_contiguous_bytes)
           = destination_address_space->convert (
-            *wave, lane_id, lowered_address_space, lowered_address);
+            *wave, lane_id, *source_address_space, source_segment_address);
       }
   }
   CATCH (AMD_DBGAPI_STATUS_ERROR_NOT_INITIALIZED,
@@ -1181,14 +1191,17 @@ amd_dbgapi_address_is_in_address_class (
         if (wave == nullptr)
           THROW (AMD_DBGAPI_STATUS_ERROR_INVALID_WAVE_ID);
 
+        auto [lowered_address_space, lowered_address]
+          = address_space->lower (wave->agent (), segment_address);
+
         if (lane_id == AMD_DBGAPI_LANE_NONE
-            && address_space->address_dependency (segment_address)
+            && lowered_address_space.address_dependency (lowered_address)
                  == AMD_DBGAPI_SEGMENT_ADDRESS_DEPENDENCE_LANE)
           THROW (AMD_DBGAPI_STATUS_ERROR_INVALID_LANE_ID);
 
         *address_class_state
-          = address_space->address_is_in_address_class (
-              *wave, lane_id, segment_address, *address_class)
+          = lowered_address_space.address_is_in_address_class (
+              *wave, lane_id, lowered_address, *address_class)
               ? AMD_DBGAPI_ADDRESS_CLASS_STATE_MEMBER
               : AMD_DBGAPI_ADDRESS_CLASS_STATE_NOT_MEMBER;
       }
@@ -1205,11 +1218,14 @@ amd_dbgapi_address_is_in_address_class (
 
 amd_dbgapi_status_t AMD_DBGAPI
 amd_dbgapi_address_dependency (
+  [[maybe_unused]] amd_dbgapi_process_id_t process_id,
+  amd_dbgapi_wave_id_t wave_id,
   amd_dbgapi_address_space_id_t address_space_id,
   amd_dbgapi_segment_address_t segment_address,
   amd_dbgapi_segment_address_dependency_t *segment_address_dependency)
 {
-  TRACE_BEGIN (param_in (address_space_id), param_in (segment_address),
+  TRACE_BEGIN (param_in (process_id), param_in (wave_id),
+               param_in (address_space_id), param_in (segment_address),
                param_in (segment_address_dependency));
   TRY
   {
@@ -1221,12 +1237,43 @@ amd_dbgapi_address_dependency (
     if (address_space == nullptr)
       THROW (AMD_DBGAPI_STATUS_ERROR_INVALID_ADDRESS_SPACE_ID);
 
-    *segment_address_dependency
-      = address_space->address_dependency (segment_address);
+    process_t *process = process_t::find (process_id);
+
+    if (process == nullptr)
+      THROW (AMD_DBGAPI_STATUS_ERROR_INVALID_PROCESS_ID);
+
+    const wave_t *wave = find (wave_id);
+
+    if (wave == nullptr)
+      {
+        if (wave_id != AMD_DBGAPI_WAVE_NONE)
+          THROW (AMD_DBGAPI_STATUS_ERROR_INVALID_WAVE_ID);
+
+        if (address_space->address_dependency (segment_address)
+            != AMD_DBGAPI_SEGMENT_ADDRESS_DEPENDENCE_PROCESS)
+          THROW (AMD_DBGAPI_STATUS_ERROR_INVALID_ARGUMENT_COMPATIBILITY);
+
+        *segment_address_dependency
+          = AMD_DBGAPI_SEGMENT_ADDRESS_DEPENDENCE_PROCESS;
+      }
+    else
+      {
+        if (!wave->architecture ().is_address_space_supported (*address_space))
+          THROW (AMD_DBGAPI_STATUS_ERROR_INVALID_ARGUMENT_COMPATIBILITY);
+
+        auto [lowered_address_space, lowered_address]
+          = address_space->lower (wave->agent (), segment_address);
+
+        *segment_address_dependency
+          = lowered_address_space.address_dependency (lowered_address);
+      }
   }
   CATCH (AMD_DBGAPI_STATUS_ERROR_NOT_INITIALIZED,
+         AMD_DBGAPI_STATUS_ERROR_INVALID_PROCESS_ID,
+         AMD_DBGAPI_STATUS_ERROR_INVALID_WAVE_ID,
          AMD_DBGAPI_STATUS_ERROR_INVALID_ADDRESS_SPACE_ID,
-         AMD_DBGAPI_STATUS_ERROR_INVALID_ARGUMENT);
+         AMD_DBGAPI_STATUS_ERROR_INVALID_ARGUMENT,
+         AMD_DBGAPI_STATUS_ERROR_INVALID_ARGUMENT_COMPATIBILITY);
   TRACE_END (make_ref (param_out (segment_address_dependency)));
 }
 
@@ -1274,25 +1321,42 @@ xfer_memory (amd_dbgapi_process_id_t process_id, amd_dbgapi_wave_id_t wave_id,
         THROW (AMD_DBGAPI_STATUS_ERROR_INVALID_LANE_ID);
     }
   else if (wave_id != AMD_DBGAPI_WAVE_NONE)
-    THROW (AMD_DBGAPI_STATUS_ERROR_INVALID_WAVE_ID);
+    {
+      /* wave is nullptr and wave_id != WAVE_NONE.  */
+      THROW (AMD_DBGAPI_STATUS_ERROR_INVALID_WAVE_ID);
+    }
   else if (lane_id != AMD_DBGAPI_LANE_NONE)
-    THROW (AMD_DBGAPI_STATUS_ERROR_INVALID_LANE_ID);
+    {
+      /* wave_id is WAVE_NONE and lane_id != LANE_NONE.  */
+      THROW (AMD_DBGAPI_STATUS_ERROR_INVALID_LANE_ID);
+    }
+  else if (address_space->address_dependency (segment_address)
+           != AMD_DBGAPI_SEGMENT_ADDRESS_DEPENDENCE_PROCESS)
+    {
+      /* wave_id is WAVE_NONE and the address depends on
+         wave/workgroup/agent.  */
+      THROW (AMD_DBGAPI_STATUS_ERROR_INVALID_ARGUMENT_COMPATIBILITY);
+    }
 
   try
     {
-      switch (address_space->address_dependency (segment_address))
+      /* lower the address space if wave is not nullptr.  */
+      auto [lowered_address_space, lowered_address]
+        = (wave != nullptr)
+            ? address_space->lower (wave->agent (), segment_address)
+            : std::make_pair (std::cref (*address_space), segment_address);
+
+      switch (lowered_address_space.address_dependency (lowered_address))
         {
         case AMD_DBGAPI_SEGMENT_ADDRESS_DEPENDENCE_PROCESS:
           *value_size = process->xfer_segment_memory (
-            *address_space, segment_address, read, write, *value_size);
+            lowered_address_space, lowered_address, read, write, *value_size);
           break;
 
         case AMD_DBGAPI_SEGMENT_ADDRESS_DEPENDENCE_WORKGROUP:
-          if (wave == nullptr)
-            THROW (AMD_DBGAPI_STATUS_ERROR_INVALID_WAVE_ID);
-
+          dbgapi_assert (wave != nullptr);
           *value_size = wave->workgroup ().xfer_segment_memory (
-            *address_space, segment_address, read, write, *value_size);
+            lowered_address_space, lowered_address, read, write, *value_size);
           break;
 
         case AMD_DBGAPI_SEGMENT_ADDRESS_DEPENDENCE_LANE:
@@ -1301,15 +1365,14 @@ xfer_memory (amd_dbgapi_process_id_t process_id, amd_dbgapi_wave_id_t wave_id,
           [[fallthrough]];
 
         case AMD_DBGAPI_SEGMENT_ADDRESS_DEPENDENCE_WAVE:
-          if (wave == nullptr)
-            THROW (AMD_DBGAPI_STATUS_ERROR_INVALID_WAVE_ID);
-
-          *value_size
-            = wave->xfer_segment_memory (*address_space, segment_address,
-                                         lane_id, read, write, *value_size);
+          dbgapi_assert (wave != nullptr);
+          *value_size = wave->xfer_segment_memory (lowered_address_space,
+                                                   lowered_address, lane_id,
+                                                   read, write, *value_size);
           break;
 
-        default:
+        case AMD_DBGAPI_SEGMENT_ADDRESS_DEPENDENCE_AGENT:
+        case AMD_DBGAPI_SEGMENT_ADDRESS_DEPENDENCE_NONE:
           dbgapi_assert_not_reached ("unsupported address dependency");
         }
     }